requirements.txt
biopython
pandas
matplotlib
